* Fixed NPC_INVINCIBLE can be dispelled. (bugreport:4046) * Fixed a compliation warning.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@14247 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/clif.c
@@ -6616,7 +6616,7 @@ int clif_guild_memberlist(struct map_session_data *sd)
 		WFIFOW(fd,c*104+16)=m->gender;
 		WFIFOW(fd,c*104+18)=m->class_;
 		WFIFOW(fd,c*104+20)=m->lv;
-		WFIFOL(fd,c*104+22)=cap_value(m->exp,0,INT_MAX);
+		WFIFOL(fd,c*104+22)=(int)cap_value(m->exp,0,INT_MAX);
 		WFIFOL(fd,c*104+26)=m->online;
 		WFIFOL(fd,c*104+30)=m->position;
 		memset(WFIFOP(fd,c*104+34),0,50);	// ƒƒ‚H
